chore: build

+++ docs/static/ui/components/Histogram/drawConstant.js
@@ -1,6 +1,7 @@
-export default`precision highp float;
+export default`precision mediump float;
 #define GLSLIFY 1
 
 void main() {
     gl_FragColor = vec4(1.0);
-}`;
+}
+`;
